+++ admin-search/dev/src/payload.config.ts
@@ -22,6 +22,7 @@ export default buildConfig({
       email: 'dev@payloadcms.com',
       password: 'test',
     },
+    meta: { titleSuffix: '- Admin Search Dev' },
     user: 'users',
   },
   collections: [