+++ alt-text/src/components/icons/ArrowRightIcon.tsx
@@ -2,18 +2,19 @@ import React from 'react'
 
 export const ArrowRightIcon: React.FC<React.SVGProps<SVGSVGElement>> = (props) => (
   <svg
-    fill="none"
+    data-testid="geist-icon"
     height="16"
-    shapeRendering="geometricPrecision"
-    stroke="currentColor"
-    strokeLinecap="round"
     strokeLinejoin="round"
-    strokeWidth="1.5"
     style={{ color: 'currentcolor' }}
-    viewBox="0 0 24 24"
+    viewBox="0 0 16 16"
     width="16"
     {...props}
   >
-    <path d="M5 12h14M12 5l7 7-7 7" />
+    <path
+      clipRule="evenodd"
+      d="M9.53 2.22L9 1.69 7.94 2.75l.53.53 3.97 3.97H1v1.5h11.44l-3.97 3.97-.53.53L9 14.31l.53-.53 5.074-5.073a1 1 0 000-1.414L9.53 2.22z"
+      fill="currentColor"
+      fillRule="evenodd"
+    />
   </svg>
 )