+++ alt-text/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- style: standardize icons to use Geist icon set (16x16 filled)
 - refactor: improve widget translations (pluralize title, simplify German translations, use i18next plural keys for bulk generate button)
 
 ## 0.4.3